refactor: remove NetworkManager class and delegate DHT operations to DHTOperations class
- Deleted the NetworkManager class from the codebase, consolidating its responsibilities into the DHTOperations class.
- Updated KademliaNode to utilize DHTOperations for store, secureStore, findValue, and gateway registration methods.
- Removed unused imports and methods related to the old NetworkManager.
- Adjusted tests to reflect the removal of supportedProtocols from gateway data.
